Extract GraphMaker to its own module.
[ci skip-jvm-tests]
